renaming HomePage
