+++ src/App.js
@@ -1,7 +1,7 @@
 import { Routes, Route, BrowserRouter } from "react-router-dom";
 import "./App.css";
 import Navbar from "./components/Navbar";
-import Home from "./pages/Home";
+import Home from "./pages/HomePage";
 import AllRestaurants from "./pages/AllRestaurants";
 import AllReservations from "./pages/AllReservations.jsx";
 import Restaurant from "./pages/Restaurant";
